requirements.txt
pytesseract
pdf2image
pypdf
ollama
google-genai
